+++ common/src/comp/skills.rs
@@ -103,6 +103,8 @@ pub enum Skill {
     Sceptre(SceptreSkill),
     UnlockGroup(SkillGroupKind),
     Roll(RollSkill),
+    Climb(ClimbSkill),
+    Swim(SwimSkill),
 }
 
 pub enum SkillError {
@@ -248,6 +250,17 @@ pub enum RollSkill {
     Cost,
     Strength,
     Duration,
+}
+
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ClimbSkill {
+    Cost,
+    Speed,
+}
+
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
+pub enum SwimSkill {
+    Speed,
 }
 
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]