+++ common/src/states/climb.rs
@@ -1,5 +1,8 @@
 use crate::{
-    comp::{CharacterState, Climb, EnergySource, InputKind, Ori, StateUpdate},
+    comp::{
+        skills::{ClimbSkill::*, Skill},
+        CharacterState, Climb, EnergySource, InputKind, Ori, StateUpdate,
+    },
     consts::GRAVITY,
     event::LocalEvent,
     states::{
@@ -11,11 +14,43 @@ use crate::{
 use serde::{Deserialize, Serialize};
 use vek::*;
 
-const HUMANOID_CLIMB_ACCEL: f32 = 24.0;
-const CLIMB_SPEED: f32 = 5.0;
+/// Separated out to condense update portions of character state
+#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct StaticData {
+    pub energy_cost: f32,
+    pub movement_speed: f32,
+}
 
-#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize, Eq, Hash)]
-pub struct Data;
+#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct Data {
+    /// Struct containing data that does not change over the course of the
+    /// character state
+    pub static_data: StaticData,
+}
+
+impl Data {
+    pub fn create_adjusted_by_skills(join_data: &JoinData) -> Self {
+        let mut data = Data::default();
+        if let Ok(Some(level)) = join_data.stats.skill_set.skill_level(Skill::Climb(Cost)) {
+            data.static_data.energy_cost *= 0.8_f32.powi(level.into());
+        }
+        if let Ok(Some(level)) = join_data.stats.skill_set.skill_level(Skill::Climb(Speed)) {
+            data.static_data.movement_speed *= 1.2_f32.powi(level.into());
+        }
+        data
+    }
+}
+
+impl Default for Data {
+    fn default() -> Self {
+        Data {
+            static_data: StaticData {
+                energy_cost: 5.0,
+                movement_speed: 5.0,
+            },
+        }
+    }
+}
 
 impl CharacterBehavior for Data {
     fn behavior(&self, data: &JoinData) -> StateUpdate {
@@ -38,19 +73,18 @@ impl CharacterBehavior for Data {
             update.character = CharacterState::Idle {};
             return update;
         };
-
         // Move player
         update.vel.0 += Vec2::broadcast(data.dt.0)
             * data.inputs.move_dir
-            * if update.vel.0.magnitude_squared() < CLIMB_SPEED.powi(2) {
-                HUMANOID_CLIMB_ACCEL
+            * if update.vel.0.magnitude_squared() < self.static_data.movement_speed.powi(2) {
+                self.static_data.movement_speed.powi(2)
             } else {
                 0.0
             };
 
         // Expend energy if climbing
         let energy_use = match climb {
-            Climb::Up => 5,
+            Climb::Up => self.static_data.energy_cost as i32,
             Climb::Down => 1,
             Climb::Hold => 1,
         };
@@ -74,8 +108,12 @@ impl CharacterBehavior for Data {
 
         // Apply Vertical Climbing Movement
         match climb {
-            Climb::Down => update.vel.0.z += data.dt.0 * (GRAVITY - HUMANOID_CLIMB_ACCEL),
-            Climb::Up => update.vel.0.z += data.dt.0 * (GRAVITY + HUMANOID_CLIMB_ACCEL),
+            Climb::Down => {
+                update.vel.0.z += data.dt.0 * (GRAVITY - self.static_data.movement_speed.powi(2))
+            },
+            Climb::Up => {
+                update.vel.0.z += data.dt.0 * (GRAVITY + self.static_data.movement_speed.powi(2))
+            },
             Climb::Hold => update.vel.0.z += data.dt.0 * GRAVITY,
         }