+++ voxygen/src/hud/diary.rs
@@ -163,6 +163,11 @@ widget_ids! {
         skill_general_roll_1,
         skill_general_roll_2,
         skill_general_roll_3,
+        skill_general_climb_0,
+        skill_general_climb_1,
+        skill_general_climb_2,
+        skill_general_swim_0,
+        skill_general_swim_1,
     }
 }
 
@@ -572,6 +577,7 @@ impl<'a> Widget for Diary<'a> {
             _ => 0,
         };
         let skills_bot_r = match sel_tab {
+            SelectedSkillTree::General => 5,
             SelectedSkillTree::Weapon(ToolKind::Sword) => 1,
             SelectedSkillTree::Weapon(ToolKind::Bow) => 1,
             _ => 0,
@@ -701,7 +707,13 @@ impl<'a> Widget for Diary<'a> {
         let art_size = [320.0, 320.0];
         match sel_tab {
             SelectedSkillTree::General => {
-                use skills::{GeneralSkill::*, RollSkill::*, SkillGroupKind::*};
+                use skills::{
+                    ClimbSkill,
+                    GeneralSkill::*,
+                    RollSkill::{self, *},
+                    SkillGroupKind::*,
+                    SwimSkill,
+                };
                 use ToolKind::*;
                 // General Combat
                 Image::new(animate_by_pulse(
@@ -964,7 +976,7 @@ impl<'a> Widget for Diary<'a> {
                 {
                     events.push(Event::UnlockSkill(skill));
                 };
-                let skill = Skill::Roll(Cost);
+                let skill = Skill::Roll(RollSkill::Cost);
                 if create_skill_button(
                     self.imgs.utility_cost_skill,
                     state.skills_bot_l[1],
@@ -1018,7 +1030,7 @@ impl<'a> Widget for Diary<'a> {
                 };
                 let skill = Skill::Roll(Duration);
                 if create_skill_button(
-                    self.imgs.utility_amount_skill,
+                    self.imgs.utility_duration_skill,
                     state.skills_bot_l[3],
                     &self.stats.skill_set,
                     skill,
@@ -1038,6 +1050,108 @@ impl<'a> Widget for Diary<'a> {
                     TEXT_COLOR,
                 )
                 .set(state.skill_general_roll_3, ui)
+                .was_clicked()
+                {
+                    events.push(Event::UnlockSkill(skill));
+                };
+                // Bottom right skills
+                Button::image(self.imgs.skill_climbing_skill)
+                    .w_h(74.0, 74.0)
+                    .mid_top_with_margin_on(state.skills_bot_r[0], 3.0)
+                    .with_tooltip(
+                        self.tooltip_manager,
+                        &self.localized_strings.get("hud.skill.climbing_title"),
+                        &self.localized_strings.get("hud.skill.climbing"),
+                        &diary_tooltip,
+                        TEXT_COLOR,
+                    )
+                    .set(state.skill_general_climb_0, ui);
+                let skill = Skill::Climb(ClimbSkill::Cost);
+                if create_skill_button(
+                    self.imgs.utility_cost_skill,
+                    state.skills_bot_r[1],
+                    &self.stats.skill_set,
+                    skill,
+                    self.fonts,
+                    &get_skill_label(skill, &self.stats.skill_set),
+                )
+                .with_tooltip(
+                    self.tooltip_manager,
+                    &self.localized_strings.get("hud.skill.climbing_cost_title"),
+                    &add_sp_cost_tooltip(
+                        &self.localized_strings.get("hud.skill.climbing_cost"),
+                        skill,
+                        &self.stats.skill_set,
+                        &self.localized_strings,
+                    ),
+                    &diary_tooltip,
+                    TEXT_COLOR,
+                )
+                .set(state.skill_general_climb_1, ui)
+                .was_clicked()
+                {
+                    events.push(Event::UnlockSkill(skill));
+                };
+                let skill = Skill::Climb(ClimbSkill::Speed);
+                if create_skill_button(
+                    self.imgs.utility_speed_skill,
+                    state.skills_bot_r[2],
+                    &self.stats.skill_set,
+                    skill,
+                    self.fonts,
+                    &get_skill_label(skill, &self.stats.skill_set),
+                )
+                .with_tooltip(
+                    self.tooltip_manager,
+                    &self.localized_strings.get("hud.skill.climbing_speed_title"),
+                    &add_sp_cost_tooltip(
+                        &self.localized_strings.get("hud.skill.climbing_speed"),
+                        skill,
+                        &self.stats.skill_set,
+                        &self.localized_strings,
+                    ),
+                    &diary_tooltip,
+                    TEXT_COLOR,
+                )
+                .set(state.skill_general_climb_2, ui)
+                .was_clicked()
+                {
+                    events.push(Event::UnlockSkill(skill));
+                };
+
+                Button::image(self.imgs.skill_swim_skill)
+                    .w_h(74.0, 74.0)
+                    .mid_top_with_margin_on(state.skills_bot_r[3], 3.0)
+                    .with_tooltip(
+                        self.tooltip_manager,
+                        &self.localized_strings.get("hud.skill.swim_title"),
+                        &self.localized_strings.get("hud.skill.swim"),
+                        &diary_tooltip,
+                        TEXT_COLOR,
+                    )
+                    .set(state.skill_general_swim_0, ui);
+                let skill = Skill::Swim(SwimSkill::Speed);
+                if create_skill_button(
+                    self.imgs.utility_speed_skill,
+                    state.skills_bot_r[4],
+                    &self.stats.skill_set,
+                    skill,
+                    self.fonts,
+                    &get_skill_label(skill, &self.stats.skill_set),
+                )
+                .with_tooltip(
+                    self.tooltip_manager,
+                    &self.localized_strings.get("hud.skill.swim_speed_title"),
+                    &add_sp_cost_tooltip(
+                        &self.localized_strings.get("hud.skill.swim_speed"),
+                        skill,
+                        &self.stats.skill_set,
+                        &self.localized_strings,
+                    ),
+                    &diary_tooltip,
+                    TEXT_COLOR,
+                )
+                .set(state.skill_general_swim_1, ui)
                 .was_clicked()
                 {
                     events.push(Event::UnlockSkill(skill));
@@ -3537,7 +3651,7 @@ impl<'a> Widget for Diary<'a> {
                 };
                 let skill = Skill::Sceptre(ADuration);
                 if create_skill_button(
-                    self.imgs.buff_speed_skill,
+                    self.imgs.buff_duration_skill,
                     state.skills_bot_l[2],
                     &self.stats.skill_set,
                     skill,